Refactor: Move theme_switcher component to correct directory
The theme_switcher component files were moved from `theme_swtcher` to `theme_switcher` to correct a typo.
