+++ examples/components/cards/theme_switcher/theme_switcher_gen.c
